Changed az encoder size from 32 to 16 bits. Removed check to see if client is available. Fixed packets buffer size for temp sensors and encoders.

+++ include/Teensy41TCP_IP.hpp
@@ -4,5 +4,5 @@
 #include <NativeEthernet.h>
 
 void SendDataToControlRoom(uint8_t *, size_t, IPAddress, uint16_t, EthernetClient);
-uint32_t calcTransitSize(int16_t, int16_t, int16_t, int16_t, int16_t,int16_t,int32_t);
-void prepairTransit(uint8_t *, uint32_t, std::queue <acc> *, std::queue <acc> *, std::queue <acc> *, std::queue <int16_t> *, std::queue <int16_t> *,std::queue <int16_t> *,std::queue <int32_t> *);
+uint32_t calcTransitSize(int16_t, int16_t, int16_t, int16_t, int16_t,int16_t,int16_t);
+void prepairTransit(uint8_t *, uint32_t, std::queue <acc> *, std::queue <acc> *, std::queue <acc> *, std::queue <int16_t> *, std::queue <int16_t> *,std::queue <int16_t> *,std::queue <int16_t> *);
