+++ include/procAzEnEvent.h
@@ -4,7 +4,7 @@
 class AzimuthEncoder
 {
 public:
-    std::queue <int32_t> buffer;
+    std::queue <int16_t> buffer;
     AzimuthEncoder();
     void init(void);
     void procAzEnEvent(void);
